Get rid of some compiler warnings; Fix chdir handling of errors
-) Some "unused parameter" warnings
-) Some "ignoring return value" are a bit trickier. They are bad coding style:
    o) the file-name.cc one returns a string with undefined contents if getcwd
       fails, which might either crash guile or lead to other bugs.
    o) the lily-parser-scheme.cc warning was also a programming error (if you
       used --output=dir/file, and dir/ had the wrong permissions, then
       lilypond would not print an error, but simply put the output file into
       the current directory without telling the user!).
       In this case (output dir explicitly requested, but not possible
       to change there), I think it's best to exit lilypond with an
       error.
-) signed/unsigned warning in glissando-engraver.cc, where we already
   had a check for negative values, so explicitly casting to unsigned
   shuts up the compiler.
-) Remove unneccessary #(set-paper-size "a6") in the regtests
-) fix bad texinfo code in function documentation

+++ lily/glissando-engraver.cc
@@ -121,7 +121,7 @@ Glissando_engraver::acknowledge_note_column (Grob_info info)
               continue;
             int n1 = robust_scm2int (scm_car (candidate), -1);
             int n2 = robust_scm2int (scm_cdr (candidate), -1);
-            if (n1 < 0 || n2 < 0 || n1 >= note_heads.size ())
+            if ((n1 < 0) || (n2 < 0) || (size_t(n1) >= note_heads.size ()))
               continue;
             note_column_1.push_back (vsize (n1));
             note_column_2.push_back (vsize (n2));
